# Allow Write Operations when Slot is in MIGRATING state