--- a/libs/cluster/CmdStrings.cs
+++ b/libs/cluster/CmdStrings.cs
@@ -61,9 +61,10 @@ namespace Garnet.cluster
         /// <summary>
         /// Simple error respone strings, i.e. these are of the form "-errorString\r\n"
         /// </summary>
-        public static ReadOnlySpan<byte> RESP_ERR_CROSSLOT => "CROSSSLOT Keys in request do not hash to the same slot"u8;
+        public static ReadOnlySpan<byte> RESP_ERR_CROSSSLOT => "CROSSSLOT Keys in request do not hash to the same slot"u8;
         public static ReadOnlySpan<byte> RESP_ERR_CLUSTERDOWN => "CLUSTERDOWN Hash slot not served"u8;
         public static ReadOnlySpan<byte> RESP_ERR_MIGRATING => "MIGRATING"u8;
+        public static ReadOnlySpan<byte> RESP_ERR_TRYAGAIN => "TRYAGAIN Multiple keys request during rehashing of slot"u8;
         public static ReadOnlySpan<byte> RESP_ERR_CREATE_SYNC_SESSION_ERROR => "PRIMARY-ERR Failed creating replica sync session task"u8;
         public static ReadOnlySpan<byte> RESP_ERR_RETRIEVE_SYNC_SESSION_ERROR => "PRIMARY-ERR Failed retrieving replica sync session"u8;
         public static ReadOnlySpan<byte> RESP_ERR_IOERR => "IOERR Migrate keys failed"u8;