--- a/libs/cluster/Server/Migration/KeyMigrateState.cs
+++ b/libs/cluster/Server/Migration/KeyMigrateState.cs
@@ -6,21 +6,26 @@ namespace Garnet.cluster
     /// <summary>
     /// Used to mark the state of key owned by single given MigrateSession during key migration
     /// </summary>
-    public enum KeyMigrateState : byte
+    public enum KeyMigrationStatus : byte
     {
         /// <summary>
-        /// In queue for migration, can serve reads and writes if key exists
+        /// Key queued for migration, can serve both read and write requests
         /// </summary>
-        PENDING,
+        QUEUED,
 
         /// <summary>
-        /// Actively being migrated, writes will be delayed
+        /// Key is migrating, can serve only read requests. Write requests are delayed (i.e spin-wait)
         /// </summary>
         MIGRATING,
 
         /// <summary>
-        /// Migrated to target node, reads and writes will be redirected through ASK
+        /// Key is being deleted, reads and writes requests are delayed (i.e. spin-wait)
         /// </summary>
-        MIGRATED
+        DELETING,
+
+        /// <summary>
+        /// If Key existed it has been migrated to target node
+        /// </summary>
+        MIGRATED,
     }
 }